Fixed RS1 Tone can load from DLC Template and Knob don't enable for edit

+++ RocksmithToolkitLib/DLCPackage/Tone/Pedal.cs
@@ -8,11 +8,11 @@ namespace RocksmithToolkitLib.DLCPackage.Tone
     public class Pedal
     {
         public string PedalKey { get; set; }
-        public Dictionary<string, float> KnobValues { get; set; }
+        public Dictionary<string, decimal> KnobValues { get; set; }
 
         public Pedal()
         {
-            KnobValues = new Dictionary<string, float>();
+            KnobValues = new Dictionary<string, decimal>();
         }
 
         public bool Any() {
